🏷️ cleanup some models
